backend/requirements.txt
fastapi
uvicorn
python-multipart
llmwhisperer-client
python-dotenv
